% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{phack_test_binomial}
\alias{phack_test_binomial}
\title{Binomial test for p-hacking}
\usage{
phack_test_binomial(
  p,
  p_min = 0.04,
  p_max = 0.05,
  open_interval = FALSE,
  min_bunch = 3
)
}
\arguments{
\item{p}{vector of p-values (should be derounded if there are rounding errors)}

\item{p_min}{lower bound of the interval used for the test}

\item{p_max}{upper bound of the interval used for the test}

\item{open_interval}{if TRUE take p-values from open interval (p_min, p_max).}

\item{min_bunch}{minimal number of elements of p that have exactly the same p-value in order to show a warning that there seems to be a rounding problem.}
}
\description{
Elliot et al. (2022) show that absent rounding errors, p-hacking and
publication bias, the density of p-values across many different tests
should be decreasing in the p-value.
}
\details{
This means if we split any interval [p_min, p_max] in the center, a
significantly higher proportion in the right half than the left half
suggests a violation of the no p-hacking and no publication bias assumption
(if there are no rounding errors or p-values are approbriately de-rounded).

This function tests this via a Binomial test.
}
