% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{phack_test_cox_shi}
\alias{phack_test_cox_shi}
\title{Cox-Shi histogram test and more general test for K-monotonicity and bounds on [p_min, p_max] interval}
\usage{
phack_test_cox_shi(
  p,
  article = NA,
  p_min = 0,
  p_max = 0.15,
  J = 30,
  K = 1,
  use_bounds = FALSE,
  min_bunch = 3
)
}
\arguments{
\item{p}{vector of p-values (make sure rounding problems are dealt with)}

\item{article}{vector of unique article ids for approbriate clustuering}

\item{J}{number of subintervals}

\item{K}{degree of K-monotonicity (see Section 4.3 in Elliot et al. 2022)}

\item{use_bounds}{use bounds or test without bounds (see Appendix A in Elliot et al. 2022)}

\item{min_bunch}{minimal number of elements of p that have exactly the same p-value in order to show a warning that there seems to be a rounding problem.}
}
\description{
For the defaults K=1 and use_bounds=FALSE we have a basic histogram test.
}
