% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{phack_test_discontinuity}
\alias{phack_test_discontinuity}
\title{Discontinuity test}
\usage{
phack_test_discontinuity(p, c, min_bunch = 3)
}
\arguments{
\item{p}{vector of p-values}

\item{c}{potential discontinuity point}

\item{min_bunch}{minimal number of elements of p that have exactly the same p-value in order to show a warning that there seems to be a rounding problem.}
}
\description{
Discontinuity test
}
