% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{phack_test_fisher}
\alias{phack_test_fisher}
\title{Fisher's test}
\usage{
phack_test_fisher(p, p_min, p_max, min_bunch = 3)
}
\arguments{
\item{p}{vector of p-values (should be derounded if there are rounding errors)}

\item{p_min}{lower bound of the interval used for the test}

\item{p_max}{upper bound of the interval used for the test}

\item{min_bunch}{minimal number of elements of p that have exactly the same p-value in order to show a warning that there seems to be a rounding problem.}
}
\description{
Similar to phack_binomial_test but using Fisher's test instead
of the binomial test.
}
\details{
For this test the half-open interval [p_min, p_max) is used.
}
