% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{phack_test_lcm}
\alias{phack_test_lcm}
\title{LCM test on [p_min, p_max]}
\usage{
phack_test_lcm(p, p_min, p_max, F_LCMsup = get.phack.F_LCMsup(), min_bunch = 3)
}
\arguments{
\item{p}{-- vector of p-values}

\item{p_min}{lower bound of the interval used for the test}

\item{p_max}{upper bound of the interval used for the test}

\item{F_LCMsup}{cdf for LCM test}

\item{min_bunch}{minimal number of elements of p that have exactly the same p-value in order to show a warning that there seems to be a rounding problem.}
}
\description{
LCM test on [p_min, p_max]
}
