% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{SimBB}
\alias{SimBB}
\title{Simulate Brownian Bridge (BB) and ||LCM(BB)-BB||}
\usage{
SimBB(M)
}
\arguments{
\item{M}{-- number of repetitions}
}
\description{
Simulate Brownian Bridge (BB) and ||LCM(BB)-BB||
}
